Using ClassGraph instead of Reflections

+++ common/src/main/java/team/_0mods/aeternus/api/client/utils/ClientUtils.kt
@@ -40,7 +40,7 @@ fun LivingEntity.render(
     guiGraphics: GuiGraphics,
     x: Float,
     y: Float,
-    scale: Int,
+    scale: Float,
     mouseX: Float,
     mouseY: Float,
 ) {
